+++ examples/plot_digits_pipe.py
@@ -57,7 +57,7 @@ estimator = GridSearchCV(pipe,
                               logistic__C=Cs))
 estimator.fit(X_digits, y_digits)
 
-plt.axvline(estimator.best_estimator_.named_steps['pca'].n_components,
+plt.axvline(estimator.best_estimator_.named_steps_['pca'].n_components,
             linestyle=':', label='n_components chosen')
 plt.legend(prop=dict(size=12))
 plt.show()